moved udp functions into tnfs_imp
